refactor: Extra fs function into util.js
